+++ lib/rubocop/cop/minitest/assert_nil.rb
@@ -3,13 +3,15 @@
 module RuboCop
   module Cop
     module Minitest
-      # This cop enforces the test to use `assert_nil`
-      # instead of using `assert_equal(nil, something)`.
+      # This cop enforces the test to use `assert_nil` instead of using
+      # `assert_equal(nil, something)` or `assert(something.nil?)`.
       #
       # @example
       #   # bad
       #   assert_equal(nil, actual)
       #   assert_equal(nil, actual, 'message')
+      #   assert(object.nil?)
+      #   assert(object.nil?, 'message')
       #
       #   # good
       #   assert_nil(actual)
@@ -17,27 +19,29 @@ module RuboCop
       #
       class AssertNil < Base
         include ArgumentRangeHelper
+        include NilAssertionHandleable
         extend AutoCorrector
 
-        MSG = 'Prefer using `assert_nil(%<arguments>s)` over ' \
-              '`assert_equal(nil, %<arguments>s)`.'
-        RESTRICT_ON_SEND = %i[assert_equal].freeze
+        ASSERTION_TYPE = 'assert'
+        RESTRICT_ON_SEND = %i[assert_equal assert].freeze
 
-        def_node_matcher :assert_equal_with_nil, <<~PATTERN
-          (send nil? :assert_equal nil $_ $...)
+        def_node_matcher :nil_assertion, <<~PATTERN
+          {
+            (send nil? :assert_equal nil $_ $...)
+            (send nil? :assert (send $_ :nil?) $...)
+          }
         PATTERN
 
         def on_send(node)
-          assert_equal_with_nil(node) do |actual, message|
-            message = message.first
-
-            arguments = [actual.source, message&.source].compact.join(', ')
-
-            add_offense(node, message: format(MSG, arguments: arguments)) do |corrector|
-              corrector.replace(node.loc.selector, 'assert_nil')
-              corrector.replace(first_and_second_arguments_range(node), actual.source)
-            end
+          nil_assertion(node) do |actual, message|
+            register_offense(node, actual, message)
           end
+        end
+
+        private
+
+        def assertion_type
+          ASSERTION_TYPE
         end
       end
     end